+++ RELEASENOTES.md
@@ -49,6 +49,8 @@
     *   Add `onSetMediaItems` callback listener to provide means to modify/set
         `MediaItem` list, starting index and position by session before setting
         onto Player ([#156](https://github.com/androidx/media/issues/156)).
+    *   Avoid double tap detection for non-Bluetooth media button events
+        ([#233](https://github.com/androidx/media/issues/233)).
 *   Metadata:
     *   Parse multiple null-separated values from ID3 frames, as permitted by
         ID3 v2.4.